requirements.txt
amqp==5.0.9
APScheduler==3.8.1
arabic-reshaper==2.1.3
asgiref==3.4.1
attrs==21.4.0
autobahn==22.1.1
Automat==20.2.0
billiard==3.6.4.0
boto3==1.20.46
botocore==1.23.46
celery==5.2.3
certifi==2021.10.8
cffi==1.15.0
channels==3.0.4
charset-normalizer==2.0.9
click==8.0.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.4
constantly==15.1.0
cryptography==36.0.1
daphne==3.0.2
defusedxml==0.7.1
Deprecated==1.2.13
dj-database-url==0.5.0
Django==4.0
django-allauth==0.47.0
django-apscheduler==0.6.0
django-celery-results==2.2.0
django-environ==0.8.1
django-storages==1.12.3
django-timezone-field==4.2.3
django-widget-tweaks==1.4.9
djangorestframework==3.13.1
future==0.18.2
gunicorn==20.1.0
html5lib==1.1
hyperlink==21.0.0
idna==3.3
incremental==21.3.0
jmespath==0.10.0
kombu==5.2.3
oauthlib==3.1.1
packaging==21.3
Pillow==8.4.0
prompt-toolkit==3.0.28
psycopg2==2.9.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
PyJWT==2.3.0
pyOpenSSL==22.0.0
pyparsing==3.0.7
PyPDF2==1.26.0
python-bidi==0.4.2
python-crontab==2.6.0
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2021.3
pytz-deprecation-shim==0.1.0.post0
redis==4.1.3
reportlab==3.6.6
requests==2.26.0
requests-oauthlib==1.3.0
s3transfer==0.5.0
service-identity==21.1.0
six==1.16.0
sqlparse==0.4.2
Twisted==22.1.0

txaio==21.2.1
typing_extensions==4.1.1
tzdata==2021.5
tzlocal==4.1
urllib3==1.26.7
vine==5.0.0
wcwidth==0.2.5
webencodings==0.5.1
whitenoise==5.3.0
wrapt==1.13.3
xhtml2pdf==0.2.5
zope.interface==5.4.0
